Made changes in the README file.
