+++ examples/JMS/src/assemble/appPackage.xml
