+++ examples/JMS/src/main/resources/META-INF/properties.xml
