+++ examples/JMS/src/test/resources/log4j.properties
