+++ examples/pom.xml
@@ -202,6 +202,7 @@
     <module>ftp</module>
     <module>s3</module>
     <module>jdbc</module>
+    <module>JMS</module>
   </modules>
 
   <dependencies>
